HUGE commit. Started to introduce "Click And Go" technology
Basically it's going to be possible to have a shortcut button on every Virtual Console
slider to quick access the common functionalities of the associated channels (gobos, colors, presets, RGB)
The selection is easy and visual.

With this commit Click&Go primary colors are available.
Also it is possible to add a bitmap to a gobo/preset channel in the Fixture Editor.
However, this is still an incomplete work

More to come :)