Fix dae1a8516914e61074344335a8eba6c42a61bb16 Added page change feature for Enttec Wing
Unit tests crashed after this commit -- when handling page buttons in applyExtraButtons(), the code didn't check
if data for sliders is actually there. The result was crash in tooShortData test.

Since with page flipping parseData() and applyExtraButtons() need the whole packet (with sliders) anyway,
I decided it's enough to check the size at start of the method once.