QT5 Plugins - proper IID, removed json files
I believe the IID are supposed to follow java package notation. i.e. reverse domain + class/interface name.

JSON files are not used at the moment, so I removed them.
The point is to get some data about the plugin without actually loading it; for plugins
we need to list the ports, and for that we need to load the plugin anyway.