Updated MidiDevice to load/save with device names
This changes the way MidiDevice settings are saved in QSettings

Instead of  /midiplugin/3/midichannel = 16
Now it is  /midiplugin/Input/US-2400/midichannel = 16
and so on.

To be backwards compatible, if loadSettings() can't find the new
settings key, it will try the old-style.
All saveSettings() values are saved in the new format.
Old-style settings are left in place to allow running new and old QLC+
versions. However, any new changes will not be written to the old-style
settings keys.