Address Tool: Allow address input by clicking the sliders of the DIP switch
Addresses can be entered by clicked the sliders of the DIP Switch and making them
look like the switch on the fixture. IMO that feature is useful as it is easier
to just copy the settings of a fixture than to try out addresses until you find
the correct one if you're not experienced with binary.
Also, this fixes a little glitch when resizing the Address tool window to a long width and
a short height, the sliders used to get disproportionaly large.
Also allow address 512 to be entered, or is there a good reason why it stopped at 511 before?