Engine: fixed MasterTimer (unix) eating 100% CPU
This is a delicate patch to react on system date/time changes when QLC+ is running.
Without this patch, when date changed, MasterTimer thread ate 100% of CPU without
the possibility to fix it on the fly.
This should fix issue number 1 on the Raspberry Pi, that retrieves the system
date from NTP, thus changing the date when QLC+ is already running.