ShowManager: major rework to introduce the ShowFunction class
It represents a QLC+ Function associated to a Show Track and displayed by a ShowItem
For backward compatibility, an ugly function is called on Doc load (temporaryDocFixup) to
migrate an "old" project into the new architecture. To be removed after a couple of releases.

Fixed also ShowEditor to correctly display audio/video information
Fixed the Show repaint to avoid loosing an audio preview