Monitor: Add simplified strobing
If capability contains "strob" or "pulse", 2D view will blink. It does not follow the real frequency,
nor the pulse characteristics. It is meant to symbolically show that the fixture is strobing.

This is "quick hack". To do it properly: move shutter enum to QLCCapability or QCLChannel,
add Shutter field to QLCCapability (optionally add frequency field), extend FixtureEditor appropriately
and update fixture definitions.