Reduced universesWritten signal CPU usage
Moved Universe m_hasChanged reset into zeroIntensityChannels,
so even if no output plugin is connected, the UI processes the minimum quantity of changes in a universe.

Basically the universesWritten signal goes into the App class and each Fixture is updated with their values.
If they change, Fixture emits a signal to who is connected.
This saves a lot of useless parsing in the Monitor, where there is a 1:1 fixture representation.