Fix random channels not correctly reset
From Qt QMutableHashIterator doc:

Only one mutable iterator can be active on a given hash at any time.
Furthermore, no changes should be done directly to the hash while the
iterator is active (as opposed to through the iterator), since this
could invalidate the iterator and lead to undefined behavior.