Plugins: improved architecture to save plugin parameters in the project XML
Plugins/ArtNet: implemented unicast universe transmission and ArtNet output universe setting

This in a massive commit that changes a lot the architecture of the QLC+ plugins.
The underlying QLCIOPlugin class now has a bit of intelligence, and it keeps track
of the association between QLC+ universes and the plugin lines.
It is also the place where plugins custom parameters are stored. When saving
the workspace XML, QLC+ now requests to each plugin's patched line the custom
parameters set by the user.

For now only MIDI and ArtNet use this new architecture.
TODO: OLA, OSC, (E1.31?) and those I might be forgetting right now