Engine: final fix to detect a Universe change (fix another monitor issue)
This should put the word 'end' to all my changes to reduce the number of
universesWritten signals emitted.
Unfortunaltey the trade-off is either more memory required (512 bytes per universe)
and a byte-to-byte copy to detect if there's been a change in a Universe.
The more channels used, the more CPU used :(

If anyone can come up with further optimization is very welcome.
For example an integer-casting of the QByteArray data should decrease
by 4 the CPU cycles. I tried it and miserably failed :'(