Migrated Windows build to MSYS2 (see wiki)
This simplify building on Windows A LOT.
It finally allows to use more recent Qt versions.
I used 5.4.2 and the MSYS2 package tree is constantly udpated.

MSYS2 allows also to produce 64bit builds, but then we'd have issues
with the SDK-based plugins (DMX USB and Velleman) Haven't tried it yet.