plugins: introducing 2 + 1 new plugins
Dummy: an empty plugin meant to be used as a skeleton when adding new plugins
It's commented as much as I could for educational (and reminding) purposes

UART: pure serial implementation of the DMX protocol. (Linux only)
As non-USB serial interfaces are no more available on PCs these days,
this plugin is meant to be used on embedded devices (e.g. Raspberry Pi)
Tested with BitWizard hat on the RPi: http://www.bitwizard.nl/shop/index.php?route=product/product&product_id=153
TODO: write the input implementation.

GPIO: implementation of GPIO input/output. (Linux only)
This is commented in plugins.pro cause it's not meant to be used on PCs.
Code is in place, but I need a hand to perform some real usage tests.
I have no proper devices to test it.