plugins/dmxusb: major rework of the plugin structure
Instead of having a single QLCFTDI class defined depending on the platform,
this new structure introduces subclasses of the 3 available drivers:
libFTDI, FTD2XX and QtSerialPort.

In this way it is possible to use multiple drivers on one platform.
Goal of this change is to use a combination of QtSerialPort and libFTDI on OSX,
where the first is used for what it can detect (FTDI and a few more) and libFTDI
for the rest.
It should also allow to support NanoDMX and Eurolite on OSX but more importantly
it should put the word END to the damn Apple driver issue, since QtSerialPort
uses the OSX native IOKit implementation.