plugins/dmxusb: enabled dual driver on OSX
This is the point I wanted to reach.
Now libftdi and QtSerialPort are both bundled.
QtSerialPort has the precedence except for OpenDMX devices, which will
always use libftdi.

All the rest is hopefully working out of the box, without disabling the
evil Apple FTDI USB driver.

This patch also makes NanoDMX and Eurolite (not tested) to work on OSX.