Input Profiles: added an option for MIDI profiles
to feedback a Note Off or a Note On with 0 velocity when value is 0.
APCMini now works out of the box.
This is basically what Jano did but moved into input profiles.
In the future, this global settings technique can be used for other
profiles specific settings, by just working on the input profile editor and
the input profile class