Removed QtXml dependency from Fixture Editor
This is quite a big change that continues the XML migration.
All the classes involed in a Fixture Definition now use QXMLStreamReader and Writer,
so we shouldn't have messed attributes anymore.

I had to almost rewrite completely the Avolite D4 loader. It was buggy and pretty
inefficient. There's a lot more that can be done in that area, such as heads and
palettes.

P.S. Test units are a blessing of this project ! :)