Input profiles: added more per-channel options
This patch does 2 things:
- use the sensitivity parameter for Encoders. Range is 1 to 20
- add an option for Buttons to simulate a full press-release sequence

The problem:
TouchOSC toggle buttons send only 255 on the first press/release and 0 on the second press/release.
QLC+ toggle works on two 255/0 sequences (all MIDI controllers do that)
So basically it was necessary to press a TouchOSC button 4 times to enable/disable a VC Button

The solution:
Turn the TouchOSC behaviour like MIDI controllers, by filtering toggle events and
simulate a full 255/0 sequence on every press and release.