plugins/ArtNet/E1.31/OSC: found a (hopefully) final solution for delayed network interfaces
QLC+ on a Raspberry Pi 3 boots in 12 seconds now !
The network interfaces, especially in DHCP, take a few more seconds to become ready.
With this change, QLC+ waits a maximum of 10 seconds if it has been requested an input/output line
that is not available in the system yet.
Shouldn't affect at all desktop systems, where the interfaces are always available when QLC+ starts.