Add Function alternateSpeed
The idea is: RGBMatrix innerSpeed differs from "normal" speed.
The "normal" speed is the "outer" speed.
The innerSpeed is the speed of each step.

Same for Chaser:
The innerSpeed is the speed for each step (in Common speed mode)
Other alternateSpeeds are available for Chaser,
these are the speeds of each individual step (in Per Step mode)

The alternateSpeed allows getting/setting these new Speeds.
This will allow the VCSpeedDial to change the speed steps of a Chaser.

The main goal of alternate speeds is: when changing the "normal" speed
of a chaser or rgbmatrix, the inner speed is not affected.