engine: reworked *nix MasterTimer to avoid gettimeofday
gettimeofday shouldn't be used, as explained in this article:
https://blog.habets.se/2010/09/gettimeofday-should-never-be-used-to-measure-time
Basically it suffers from system time changes, while what we need is a monotonic clock.
It was needed to adapt also the OSX code, to get an equivalent monotonic clock with nanoseconds precision.
This also improved the beat clock elapsed check. The Windows code needs to be tested.
This whole change needs to be tested on a Raspberry Pi as well.
It is not clear instead if the code block I left in #if 0 is really needed.
It would drain a lot of CPU and most likely one nanosleep should do the job.