engine: get rid of every QTime usage, and use QElapsedTimer
Again, QTime is subject to system time changes and should never be used for ms precise measurements.
Instead, we need a reliable and monotonic clock, which is provided by QElapsedTimer on every platform.
Plus, it provides also nanosecond elapsed time, which can be useful to increase the precision even further.