qmlui: major rework of the context detach mechanism
Basically a QML Window/ApplicationWindow creates a modal window, as a child of the parent window.
The new window is not separated from the parent window in the OS task bar, so when minimizing the
parent window it is impossible to recover the child window. Wicked.
I couldn't find a sense to QML windows at all actualy...
This rework, gives all the responsibility to the ContextManager, and detached windows are now
full QuickView instances, exactly like the main window. VC pages are treated as such too.