engine: fix composition of fixture definition absolute path
During tests, FIXTUREDIR is set to the production one (/usr/share/qlcplus)
but FixtureMap.xml is loaded from the source tree and definitions are not installed.
Therefore it is necessary to store the base path of the map to understand where definitions
are located. This is to avoid storing absolute paths in each definition.