engine/ui: reworked Chaser next/previous/stop/set methods
Unified into a new setAction method that allows to specify many more details
(such as intensity or fade mode) before performing an action on a Chaser step.
In this way, VCCueList has more control over steps blending and intensity adjustments
that are now applied when a step is started, and not ticks later.