engine: introduce a Function flag to preserve attributes
ATM this is used only for Sequences, since they stop/start the same Function over and over again.
The outcome was that postRun (therefore attributes reset) was called AFTER
requestAttributeOverride() + start(). With this flag, and knowing that a Chaser will use always
the same attribute ID, intensity can be preserved and applied at the right time.