Introducing DMX USB Open RX Widget.
This widget transforms an Open TX widget in an input DMX receiver only.

It use a FTDI RS232 chip and directly reads DMX frame it can catch.

FTDI RS232 chips can send and receive signals. As a DMX widget, they
are usually built with a female DMX socket, since it's the main usage.

Using a male/male adaptator and plugging in the widget in a DMX source, it
can receive the signal (instead of transmitting it) and forward it to
the host computer.

It has been tested with another DMX Open TX FTDI device ran by QLC. It
works well and is very stable (if the computer hardware is too).

It has also been tested with a Stairville DDC-6 LCD very cheap
ligthing controller. This controller vaguely speaks DMX,
with 250 Hz frame rate frequency (!) and frames of 29 channels. This
pluggin handles this case, trying to manage bogus frames.

As a result, I can drive 6 sliders (or other virtual widgets) with
the Stairville Controller using only another cheap DMX USB device.