Add a function to draw a blinder flash on an RGB matrix
The blinder has a center and a reflector and at the end flashes the matrix fully.
It can be configured to draw multiple lights on a larger matrix.
Use it in single-shot mode, reverse or in a sequence with the target color in the middle to keep the blinder on for the configured time.