Account for channel value modifed with VC widget page value when using MIDI feedback for OMNI channel.
VC frame pages encode their value in the channel value sent to the plugin, when using the MIDI OMNI
channel, this gets embedded in the bits which define the MIDI channel to send a feedback message on.
Take off the embedded page data to get the correct MIDI channel for sending feedback.

Add some simple test cases to validate the change.