RPMs also contain .so files
Build would fail because nothing picked up the .so files.

The %files pattern already matched .so.1 and .so.1.0 but not .so .

The .so files should in principle be in separate dev packages or be
deleted (together with .so.1 and .so.1.0) ... but they are just small
symlinks, so let's keep them all.