Access plugins by name and not by index
This should address what reported in #1150 and repeatedly on the forums.
However, the issue cannot be solved in all the cases:
for example HID and MIDI don't include a unique identifier (or a reliable one)
Two identical MIDI controllers cannot be distinguished at plugging time.
This is backward compatible for old projects.
If no device name match is found (e.g. IP address change), the code will fallback to the old index method.