Refactor devtool.js to be independent from other scripts.
The goal of this commit is the separation of otherwise shadowed
variables into a devtool specific object.

Also, update the stepCount if parameters are changed because this
reflects the QLC behaviour and is used in ome scripts (e.g. strobe.js).

Also, initialize the width and height before using them in script
calculations (put the initialization the required steps before).

Also, add a standard template in devtool.js to silence warnings on test execution.