Do not load RGBscript files which are not script files (#1289)
* Do not load RGBscript files which are not script files (Makefiles, backup files, directories, etc.)

* Remove filter for devtool.js as it is not packed (but only tested).

* Add a test case running all available rgbScripts to find issues within the scripts.

* Refactor devtool.js to be independent from other scripts.

The goal of this commit is the separation of otherwise shadowed
variables into a devtool specific object.

Also, update the stepCount if parameters are changed because this
reflects the QLC behaviour and is used in ome scripts (e.g. strobe.js).

Also, initialize the width and height before using them in script
calculations (put the initialization the required steps before).

Also, add a standard template in devtool.js to silence warnings on test execution.

* Remove unused variables.

* Create a better match between name and variable.

This is now also covered by the devtool.js update to use the read
function instead of accessing the variables directly. Nevertleless, this
is now more consistent.

* Fix multiple problems in the lines.js script.

- Keep the R, G and B values in the range of 0-255
- Keep the array access indexes in the range of the available array
  sizes.
- Add robustness to the setter functions which otherwise lead to a
  failing QJScript evaluation for this script.

* Hide a local variable.

* Check if each script has a name set.

* Remove devtool.js template implementation. The test now shows a warning.

* Fix typo in linesVariability variable name

* Add more checks for valid parameter definition and implementation

- Check default values (in range)
- Check set values are taken over towards get values
- Check list values are defined in list
- Check list definitions do not contain duplicates

* Return the correct value (cleanup copy & paste error in script)

* Return the same value as the value that is set.

Remove inconsistency between setter and getter.

* Add a test that the RGB value range is respected.

This is an error otherwise detected only in devtools.html when the
browser freezes when the RGB values exceed 255.
Scripts should take care to define a clean behavior when calculating the
colors.

* Make the test for RGB color overrun effective.

- QColor(Qt::red).rgb() returns 0xffff0000 due to the alpha channel, but
  0x00ff0000 is needed to test against overruns.
- Test the whole map for the pixels (by using the mapSize defined in the
  test)