Prevent color tool being covered up in fullscreen
Specify the parent of the QColorDialog when creating it for the
color tool.  This ensures that the dialog remains on top of the
main window.  Without this fix, the dialog can get covered up in
fullscreen mode under Ubuntu MATE and probably other Linux distributions.
Once the dialog is buried, recovery is impossible and the application
must be killed from a terminal window.