Fix off-by-1 issue with slider<->step# calculations
When the Step # changed, or the SliderValue changed, in Steps mode, the calculation to convert step# to value or vice versa had some off-by-one issues. The main visible symptom was that the slider position didn't always update when clicking the PreviousCue button (in a 6-step chaser, it would update the slider position every-other-click). Also, if moving the slider position using up/down arrows, the step # didn't change at the exact same slider positions as were jumped to using the Next-Cue button.

In v5 this is perhaps not necessary, because the slider position doesn't update if the Current Step changes; but if code is ever added to set the slider position on current step changing the same fix for the same reason is applicable here too.