CMake: Set LIBSSUBDIR to install shared libraries in a sub directory
On Linux, /usr/lib/*.so is reserved for libraries that are shared to be
used by other applications. They must thus have careful versioning and
development headers etc.

The qlcplus .so libraries are pretty much application internals. They
are not versioned, and are only used by qlcplus and the fixture editor
(and tests). They do not belong in the public in /usr/lib .

Instead, make it so that if building with

        cmake -DLIBSSUBDIR:STRING=/qlcplus

LIBSDIR will be set to /usr/lib/qlcplus , and both binaries will get
RPATH set to look there.

Also, with LIBSDIR already being specific to qlcplus, PLUGINDIR is just
set to the /plugins sub directory.