CMake: Introduce LIBSSUBDIR to install shared libraries in a sub directory
On Linux, /usr/lib/*.so is reserved for libraries that are shared to be
used by other applications. Such shared libraries must thus have careful
versioning and development headers etc.

The qlcplus .so libraries are pretty much application internals. They
are not versioned, and are only used by qlcplus and the fixture editor
(and tests). They do not belong in the public namespace in /usr/lib .

Static linking could be a good option without much overhead if we only
had the 2 binaries. But it is not a good option when we also consider
the test binaries.

Instead, make it so that if building with

        cmake -DLIBSSUBDIR:STRING=/qlcplus

the LIBSDIR default value of /usr/lib will be augmented to
/usr/lib/qlcplus , and the binaries will get RPATH set to look there.

Also, with LIBSDIR already being specific to qlcplus and a specific
build of it, there is no need to install plugins in a deep path like
/usr/lib/qlcplus/qt6/plugins/qlcplus . Instead, just set PLUGINDIR to
the LIBSDIR/plugins , like /usr/lib/qlcplus/plugins .

If not setting LIBSSUBDIR, everything will be as before.